refactor: split the feature flag tests into independent files
